changing to unsigned long int

+++ bit_manipulation/5-flip_bits.c
@@ -8,8 +8,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	int count = 0;
-	long int xor = n ^ m;
+	unsigned long int count = 0;
+	unsigned long int xor = n ^ m;
 
 	while (xor > 0)
 	{
